Move lzma.ts to xz.ts to avoid confusion
Change-type: patch
