+++ lib/source-destination/index.ts
@@ -23,7 +23,6 @@ export * from './driverless';
 export * from './gzip';
 export * from './file';
 export * from './http';
-export * from './lzma';
 export * from './metadata';
 export * from './multi-destination';
 export * from './progress';
@@ -31,4 +30,5 @@ export * from './balena-s3-source';
 export * from './source-destination';
 export * from './source-source';
 export * from './usbboot';
+export * from './xz';
 export * from './zip';
